Launchpad service (#7)
Launchpad service is a service that can run multiple services and
execute them sequentially. It's convienient when you have multiple
services that need to be run in a specific order or the services
aren't heavy and you don't want to use multiprocessing.
Also it simplify the configuration of the services.

Signed-off-by: Anton Kremenetsky <anton.kremenetsky@gmail.com>